#3 refactor: change request dto's access modifier to pakage
- create record pakage
- delete dto pakage
