+++ monitoring/src/main/java/monitoring/api/record/PutUpdateRecordResponseDto.java
@@ -1,4 +1,4 @@
-package monitoring.api.dto;
+package monitoring.api.record;
 
 import lombok.*;
