+++ monitoring/src/test/java/monitoring/api/record/RecordsApiControllerTest.java
@@ -1,8 +1,7 @@
-package monitoring.api;
+package monitoring.api.record;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
-import monitoring.api.dto.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
